Further tinkering with the JavaScript
I refined the ability of the script that handles the table of contents to detect whether it should have a "Back to Index" link. The previous logic of checking the file name for whether it was index.html doesn't cover cases where the book is renamed, as Spring Data does with its books. So I added a check for the documet type. A type of "book" indicates a stand-alone document, which shouldn't have hthe link.