Fix Stage and Promote release process
This commit contains a few fixes for the CI release process.

The base CI image now installs the `jq` binary as it's
required by the promotion process.

The staging script is fixed to use the correct `sed` syntax
for this Linux container, and only consider Gradle as a build
system. This also changes the staging script to correctly
choose the release version depending on the release type.
This commit also adds a `RELEASE_TYPE` parameter to the state
task.

The project build version is moved to a `gradle.properties`
to align with the stage+promote script.

The promote release script is now executable.