Apply 100% width to stretch class
Previously, Asciidoctor used spread for a full-width table but this
changed [1] to stretch in 1.5.7. As a result, a table that should be
full-width is not when using spring.css with Asciidoctor 1.5.7+.

This commit updates the CSS to apply 100% width to stretch. This is
done in addition to spread to avoid breaking backwards compatibility.

[1] https://github.com/asciidoctor/asciidoctor/issues/2589