Add feedback (#21)
* Adding a feedback widget for the reference docs

I created the first draft of a feedback widget for the reference docs.

* Added a project name field

We'll need to be able to filter by project name, so I added a field for the project name.

* Added a formatted date

For the sake of display in the viewer application, I added a formatted date field.

* Adding the necessary JavaScript libraries. Damien Vitrac caught that they were missing. Thanks, Damien.

* Adding more items that Damien discovered I had forgotten. Thanks, Damien.

* Adding more changes that Damien discovered were missing. In toc.js this time. Thanks, Damien.

* Adding the images this time.

* Review and optimisations

* Added the ability to turn off the feedback widget

The feedback widget does not appear unless an element with an ID of 'pageId' exists in the asciidoc file. The feedback widget needs an element with that ID, so that makes a sensible switch.

* Cleaning up some unneeded bits

The work Damien did made some things (jquery stuff) unnecessary. I also fixed a grammatical error.

* CSS cleanup

Damien made feedback.css be unnecessary, but that meant moving a CSS rule to a SASS file. I took care of that.

* A few more changes as the result of testing

I did some more testing (trying to be thorough) and caught a few more things. I made changes accordingly.

Co-authored-by: Damien Vitrac <oodamien@users.noreply.github.com>